Fixed cr_clarity skill clickable when it would not work
The clarity skill kept being available until 100% fatigue is reached, making it confusing on when it would actually work. To fix this, I centralized the stamina calculation logic in one function and status effects in another function. The skill can now tell if the actor would benefit from the skill, also disabling it if the user is under the listed status effect.

Changed the skill hiding logic - it is now always visible on use with a new "will activate next turn" message.

Tested the skill in battle, it correctly disables itself at high fatigue levels. If the user is over-fatigued at the turn start, it is possible to see 9/11 action points, but that correctly resets to 10 next turn if fatigue levels drop.